backend/requirements.txt
# Django and Web Framework
Django==5.0.1
djangorestframework==3.14.0
django-cors-headers==4.3.1
drf-spectacular==0.27.0
channels==4.0.0
channels-redis==4.1.0
daphne==4.0.0

# Authentication
mozilla-django-oidc==4.0.1
PyJWT==2.8.0
cryptography==42.0.0

# Database
psycopg2-binary==2.9.9

# Caching and Message Queue
redis==5.0.1
django-redis==5.4.0
celery==5.3.6
pika==1.3.2

# HTTP Client
httpx==0.26.0

# Monitoring and Metrics
prometheus-client==0.19.0
opentelemetry-api==1.22.0
opentelemetry-sdk==1.22.0

# ML Integration
numpy==1.26.3
pandas==2.2.0
scikit-learn==1.4.0
joblib==1.3.2

# Utilities
python-dotenv==1.0.0
tenacity==8.2.3

# Production Server
gunicorn==21.2.0
uvicorn[standard]==0.27.0

# Development
pytest==7.4.4
pytest-django==4.7.0
pytest-asyncio==0.23.3
black==24.1.1
flake8==7.0.0

